Introduced a new credentialBone
It's values are never read from the datastore, but stored there un-hashed.
The current use-case is to prevent leaking the otp-key to the user but can be
used for any information the should be stored but not retrevied by the user.
(It's value can still be read using db.Get()) on that entry or changing this
bone to stringBone on the model)