Speedup processing relations in skels by reusing a global refSkel / usingSkel class
Until now, for each referenced entry in any relation a new refSkel class was beeing created and assigned
to valuesCache[boneName]["dest"]. Now, we're reusing a global refSkel class and assign only its valuesCache
to our valuesCache[boneName]["dest"].

Warning: Break threadsafety in it's current form.

We'll have to decide if we want to support it and adapt accordingly.