Hotfix for PR #32
This is a serious hotfix for a problem that only comes up when using
db.KEY_SPECIAL_PROPERTY in the filter, in this case, setting the filter
property in line 483 raises an exception that a db.Key property is required,
rather than a string.

This hotfix only converts the db.Key into str for further filter if
db.KEY_SPECIAL_PROPERTY is not explicitly filtered.