Fixing relationalBone.refresh()
There was a problem with the removed condition, which comes up when there is a multi-language bone named f.e. "bonename", whose values are stored into the entity as "bonename.de" and "bonename.en".

In this case, the key, just "bonename" (which is the key in this loop) was not found in newValues, and the unserialize() function was not triggered.

In the end, changed values where not updated correctly inside an updateRelations run.