Entries fetched with db.GetOrInsert and written with db.Put will now correctly flush the memcache
This bug can cause consinstency issues with the datastore if an entry is fetched with
GetOrInsert and then later modified and written back to the datastore with db.Put.
In this case, the old version of this entry is not beeing removed from the memcache
and thus returned in subsequent db.Get() calls. This does not happen if the entry is exclusivly read
with GetOrInsert becorse GetOrInsert will run in a transaction (in which the memcache is never accessed).