Implemented transactions
We now support transactions on firestore with an arbitrary sequence of Gets/Queries/Puts/Deletes. If we encounter an change to the data we just record it as a pending change and ensure that
subsequent gets/queries return a view on the data that is consistent with what firestore would have returned if the changes had already be applied.