Added test-code for queries that run inside an transaction
For now, we record and re-run the last queries run from inside the transaction again after we applied the transaction. It must yield the same result as we constructed inside the transaction
(where we might have to fix the results to include pending updates) - unless it gets altered by a different request between our commit and our fetch. So it might fail due to false positives.