Restored relational filtering; efficency improvements for updateRelationTask, removed utils.normalizeKey
The major change are the improvements for updateRelation. We now list the foreign keys in viur-relations and check in each skel.toDB which database fileds have actually changed.
This way we don't have to update objects that only have copies of keys that did not change in that write (which will probably be the majority in most cases)