Introduced the concept of customDatabaseAdapters for skeletons
These provide a standarized API to hook skel.toDB()/delete calls intented to be used to mirror data stored in firestore
to another database system - which then can be used to run queries that are not supported on firestore or provide fulltext search capabilities.
An example implementation is provided in customdatabases.py which mimics the old rudimentary fulltext search using viur_tags.