Strip username from userpassword/login parameters
It already happened twice in two ViUR projects, that users reported a login problem which was obviously heavy to find. The reason is: There was a space character behind the user's name. In one case, this was inserted automatically by the operating system which was Android and an android keyboard app. To avoid this problem in future, a strip() function is applied onto the username to make it clean.