Create jinja template for relational bone.
The goal is to create a jinja template for the relational bone where it
should be possible to select a dataset from the linked skeleton. This
template is currently not available in the viur-core.

To overcome the limitation of 100 datasets per request a combination of
two nested for loops is used.

The name value of the subordinated skeleton represents the whole dataset.

The limit of datasets is set to 1000 to prevent overloading the client
and a disruptive delay.