Fix for bcc and cc key error when trying to send an email
This cannot have been tested on a live system. When passing a list of strings for bcc and/or cc, the core throws an uncaught exception because the keys are missing from the dataDict in the deliverEmail method. This fix is probably not elegant and may be improved by someone who knows what they are doing.