Added logic to import blobs that have been copied client-side
Some (very old projects) don't have their blobstore exposed in the cloud store yet, so these blobs have to be copied client-side.
The advance is that we don't have to resolve blob-store dlkeys to cloud-storage files, so we don't need a mapping service to be depoyed to the old project we're importing from.