Set parentnode and parentrepo before fromClient() (#402)
This change sets the readOnly-bones parentnode and parentrepo of a TreeSkel before the fromClient call. The reason for this is, that in a specific use-case, a unique value is build, which may only be unique per repository. This construction was made in a custom fromClient function implemented on the Skeleton, which requires the parentrepo to be known.