Fix docs logo and improve maintainability of CSS
The logo path wasn't correct. It was static which doesn't works on path like .../doc_tutorial/basic/index.html and resulted in a 404.
Therefore revert some changes from 3c0583dd to the preset and set the path correctly.

For maintainability un-minify the CSS and adjust the CSS for the wrapping `<a/>`.
Generate the theme.min.css for example with:
```
npx uglifycss rtdtemplate/static/css/theme.css --output rtdtemplate/static/css/theme.min.css
```