Update viur.py
Untested, have to test this myself first and will report. Since right now it is not possible to pass a custom filename for a downloadable file, I would think this should work. also the filename parameter is apparently set by default, but it's not working for me. have to investigate why.