setuptools' install_requires from requiremens.txt
This draft reads the `install_requires` configuration directly from the requirements.txt and removes the need for a double requirement maintainance from both requirements.txt and previously setup.cfg.

This attempt is the result of the discussion in https://github.com/pypa/setuptools/discussions/3433.

This branch has been published as viur-core==3.2.0-rc4 for testing purposes.