First draft with Exposed class
This is an entire rework of the previous code, and uses a special Exposed-class to wrap any exposed function or method. It is the first step to generally restructure the "hacky" solutions from the VIUR module routing and turning it into modern concepts.

There is still much stuff to do, but it's a first attempt. This state runs with an existing project, not all edge cases where tested yet.

The SEO-routing previously done by exposed shall be turned into a separate @seo decorator, with a compatiblity fallback.