refactor: Refactor `Router` & collect context-variables to `current.request.get().context` (#842)
* Initial commit, refactor of BrowseHandler to Request

This is a first attempt to refactor this module without doing too big changes.

* Filter contexts into separate Request.contexts

Contexts are already used in several ViUR projects to turn requests into a specific data environment. Context-parameters are preceded by an "@". This can be changed to a dynamic or customizable version in a next release.

This behavior is wanted:

- Request.kwargs contains all kwargs (because of backward compatibility)
- caller(kwargs) is a filtered kwargs without contexts
- Request.contexts contains only contexts

* Renamed `Request` into `Router`