fix: Always set `Secure` mode for session cookie
According to https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie#secure, and in combination with SameSite=none, Secure must be set, and is ignored on localhost since Chrome 89 and Firefox 75.

The disabling of `Secure` on local development server is therefore source for blocked cookies in combination with `SameSite`, and shall be removed.