fix: @property in module causes problems during warmup
This fixes #933 while the fix #952 doesn't work.
`hasattr(self, key)` is implemented as "do a `getattr` if this doesn't
raise an `AttributeError`, the attribute must exist".
So the property will be accessed anyway, which is a problem if it
contains code which doesn't work during warmup (because there is not
request), like accessing `current.session.get()["my_value"]` (it's
`None`).