fix: Finetuning `ViurTagsSearchAdapter` defaults (#1010)
This pull request modifies `ViurTagsSearchAdapter` to the following
defaults:

- `min_length` to 2, as some names and words are already made of 2
characters
- `max_length` to 50, as longest non-fictive words in the world are
having at least 45 characters. This is
"pneumonoultramicroscopicsilicovolcanoconiosis" in English (45 chars)
and "Rechtsschutzversicherungsgesellschaften" (39 chars) in German. The
german fictional word
"Donaudampfschifffahrtselektrizitätenhauptbetriebswerkbauunterbeamtengesellschaft"
has 80 chars, but we ignore this here.
- `substring_matching` should default to False, as all of our current
software projects, substring_matching=True as its implemented here is
not wanted or is wanted to be changed.