feat: Collect modules recursivly in `vi/config` (#995)
The viur-shop has it's modules as submodules under a main `shop`. As you
can see
[here](https://github.com/viur-framework/viur-shop/blob/191b6a56e764a709170bcd1e3f1e27f1e3fa4c1b/src/viur/shop/shop.py#L45-L53).

So for example we have this path `/shop/address/list`.
It's perfectly route-able, but the vi doesn't know it because submodules
are not collected. This PR adds a recursive walk through all the modules
and collects them in `/vi/config`.