fix: Improve `List.default_order` to respect languages config
The current default_order caused, that entries didn't show up when there was a name-bone which had a languages-setting, so it needed to be manually corrected to e.g. `name.de`. This PR fixes the behavior to either use the current or the first language. Its still possible to redefine default_order in any style like befor.