feat: Implement `fromClient(amend=True)` feature (#1150)
Implements the `fromClient()` amend-flag as needed, to only update parts
of a skeleton that where provided during an edit-action. Any other
action requires the full skeleton, as before.

This change also fixes the need for `isPostRequest` checks on the
particular actions, as this partly omitted or done after parsing all
bones, which is a performance issue.