feat-fix: Wrap `descr` and `params.category` in `translate` object for auto translating (#1223)
Bone descr aren't translated at the json renderer. The
`translate` object was actually never created. It would work if created
manually and pass it at `descr` argument, but this is not matching to
the type hint. This PR fixes these issues.

For reference, there are some notes which could expect a translate
object:

https://github.com/viur-framework/viur-core/blob/104ac87cd92655a7aa084e182c7501dc42a5dcab/src/viur/core/bones/base.py#L168

https://github.com/viur-framework/viur-core/blob/104ac87cd92655a7aa084e182c7501dc42a5dcab/src/viur/core/bones/base.py#L1272