refactor: Improved `RelationalConsistency.PreventDeletion` validation (#1244)
You can set `consistency` on a `RelationalBone` to
`RelationalConsistency.PreventDeletion` to disallow the deletion of the
referenced Skeleton (if there is a relation from this bone to it).

This works fine as long a Skeleton is referenced by only few other
Skeleton. But if thousands of Skeletons has a reference to the same
Skeleton (10000:1 relation) the `viur_incomming_relational_locks` list
has to hold 10000 keys, which exceeds the size limit of the entity.
Furthermore the transaction limit would be exceeded during writing on
10000 entites the `"{bonename}_outgoingRelationalLocks"` property too.
We cannot carry a list, neither in the one way nor in the other. (Yes,
okay, theoretically you could write _outgoingRelationalLocks outside the
transaction, maybe in a task. But even here it can be dangerous if a
skeleton contains many (multiple) `RelationalBone`. Even if the number
of outgoing ones will probably be smaller than the number of incoming
ones. But this is still an insane overhead).


That's why I thought about what I could use to solve this. And in my
opinion, we can simply use what we already have: The `viur-relations`
entities. The `consistency` is already stored in this.
So I have removed everything that relates to the maintenance of these
lock lists, ensured that these lists will no longer be present in the
datastore entites in future and replaced the "can delete" check with a
query on the `viur-relations`.
In my tests this works very well and we should even save some runtime
and memory.

---------

Co-authored-by: Jan Max Meyer <jmm@phorward.de>