fix: `Skeleton.toDB()`s `is_add` determined wrong
This is a hotfix for the problem that `is_add` is being set based on if the key is set or not. Furthermore, it is determined twice, so that the "magic"-bullshit now was moved into the transaction. It is a fix for PR #973 but previous implementation was also wrong.