feat: `FileBone(public=True)` for public files (#1241)
This PR adds public files. Public files are stored in an extra bucket
that follows this name: public-dot-{__PROJECT_ID}
The bucket must be created with ACL rights.

With getUploadUrl, the parameter public is used to decide whether it is
a public file or not. FileBone(public=True) can only refer to a public file

Public files again can deliver a servingurl, and this PR adds facilities to route these trought the ViUR system.

---------

Co-authored-by: Andreas H. Kelch <>
Co-authored-by: Jan Max Meyer <jmm@phorward.de>
Co-authored-by: agudermann <47318461+ArneGudermann@users.noreply.github.com>
Co-authored-by: Sven Eberth <se@mausbrand.de>