feat: `Skeleton.update()` transactional read/write
Drafting new `Skeleton.update()` function based on `viur.toolkit.db.set_status()`.
This function replaces the existing Skeleton.update() which should provide a dict-like behavior. For this case, the `|`- and `|=`-operators should be used. Because Skelton/SkeletonInstance is not a dict, it should provide correct CRUD functinality.

This pull request is part of #630 and a follow-up on #1264, which should be merged first.