feat: `render.render()` generalized action skel rendering (#1270)
For now, this PR only introduces a generalized function called
`render()` to every renderer which simply takes an action and a skel,
and calls the internal rendering function.

Currently, this function is only a placeholder to have nice rendering
calls for action skels. Later, this function should generally render any
answer to a render in a clean and well-structured way.