fix: Handle `RefSkel`s in `unserialize_compute` differently (#1295)
This PR fix the following bug:

If you have a relation in skeleton A to a skeleton B that has a compute
bone with a lifetime. If you then call the skeleton A but the lifetime
of the bone from the skeleton B is not included in the call because it
is not in the RefKeys.