fix: `UserSkel.__new__()` cannot be subSkel'ed (#1296)
This hotfixes the bug that a skeleton inherited from UserSkel cannot be
subSkel'ed, as an exception `TypeError: UserSkel.__new__() got an
unexpected keyword argument 'bones'` occurs.

The above exception is from viur-core 3.7, but it will raise a similar
exception in viur-core 3.6. I found this bug by porting a project from
viur-core 3.5 directly to viur-core 3.7.