fix: `Skeleton.subskel()` and `SkeletonInstance.clone()` (#1297)
Due the changes in #1259, the cloning behavior was broken.

This pull requests fixes the following:

1. Any subskel'ed SkeletonInstance globally modified bones, as they
where flagged as "cloned" but wheren't.
2. `SkeletonInstance.clone()` caused a sub-skel lose its boneMap, so it
contained all bones again.
3. Renamed `full_clone` into just `clone` to make it more familiar with
what it does. A `Skeleton.subskel(clone=True)` does now return a
standalone clone of the skeleton which can be modified.
4. Deprecated `Skeleton.subSkel()` (camel-case) now accepts old
`fullClone`-parameter again to stay backward compatible.