fix: Remove check if logged-in in `UserPassword.login()` (#1310)
Recent use cases have shown that the current logic is confusing,
especially because it is only implemented in UserPassword.

If a user logs in with GoogleLogin but does not have sufficient rights,
they are redirected back to the start page or the login page. When
attempting to log in via UserPassword, the login attempt is immediately
acknowledged with loginSucceeded(), and it is pretended that the login
was successful - although no credentials were checked.

This is confusing and inconsistent. Removing this check allows to just
login into a fresh session with the given credentials, and everything is
more convenient.