feat: Public translations interface (#1323)
This interface was already implemented in some projects, and was now
cleaned and standardized. It allows to set translate()-objects to
public, so that an unauthorized client can request them without further
authentication.

The function /json/_translate/get_public was made available to dump
translations.

Example calls:

- `/json/_translation/get_public` get public translations for current
language
- `/json/_translation/get_public?languages=en` for english translations
- `/json/_translation/get_public?languages=en&languages=de` for english
and german translations
- `/json/_translation/get_public?languages=*` for all available
languages