feat: Provide a generic `Tree.dump()` function
In many use-cases, a tree must be fetched for e.g. a workflow, a menu or similar. This helper function directly dumps a hierarchy/tree to a given level and with a given limit as JSON.

Its better, to provide a simple function for this, instead of copy a bad function multiple times in several projects.