fix: determine a better path for a new `TranslateSkel`
The problem was on the one hand that the path was first determined in `__str__`, which means that it is not the position where the translate object was created, but where it is converted to a string (this can also be in the renderer).
I have therefore moved this to __init__.

Furthermore, the stack loaded in `walk_stack()` skipped too many frames (we actually need the first frame directly), so this couldn't work either -- except for Jinja. Now we load the stack explicitly from the top.