fix: a `not caseSesitive` bone should lock the lower value
The idea of _caseINSensitive_ bones is that the upper/lower case
does not matter and `"Test"`, `"test"`, and `"TEST"` are equivalent.

If you now have a unique contstraint `UniqueLockMethod.SameValue`,
it is still possible to create the three entries.

---

Alternatively, we could of course also introduce `UniqueLockMethod.EqualValue`,
which controls this even more precisely.