fix: Remove overwriting `action` from `@deprecated` decorator (#1389)
The deprecations warnings on third party packages
(which are still in compatibility mode with 3.6) are getting annoying.
Therefore I have set up the following warning filter in my project:

```py
    warnings.filterwarnings("ignore", category=DeprecationWarning, module=r"viur\.(shop|toolkit).*",
                            message="'clonedBoneMap' was renamed into 'bone_map'")
    warnings.filterwarnings("ignore", category=DeprecationWarning, module=r"viur\.(shop|toolkit).*",
                            message=r"Call to deprecated class method (subSkel|fromDB|toDB)")
```

Unfortunately it does not work, because the deprecations in the
`skeleton` module are explicitly called with an `action `parameter. This
in turn means that these warnings are executed in a separate context
with exactly this action, all other filter conditions are overwritten or
ignored.

I therefore suggest setting the general filter in the `__init__` to
`"once"`.

The module filter was wrong at this point anyway. The `module` refers to
the module that calls the deprecated method,
not the one that triggers the warning (my fault).

This general filter can easily be customized from a relaxed `"ignore"`
to an annoying `"always"` in the project.

Alternatively, the `viur-core` could also use its own
`DeprecationWarning` subclass so that these warnings could be filtered
even more finely using the class filter.