feat: `RebuildSearchIndex` with a specific key
This is a refactoring and improvement of the RebuildSearchIndex task.

Sometimes it is necessary to refresh a single entry with a specific key, especially because the `refresh()`-logic is being used. Instead of adding a new task for this, this pull requests adds the ability to either update all kinds, all entries of a specific kind or a single entry with a given key.

This PR also refactors the naming, because "kindnames" where referred as "modules" before.