chore: Enforce config strict mode (#1411)
The new config type has now been established in the core for two
versions.
In order to force all projects that are still running in compatibility
mode to change over, I think it is a good opportunity to activate strict
mode by default for 3.8. If necessary, it can of course still be
explicitly deactivated.