fix: `Skeleton.write()` returns full skeleton (#1421)
Fixes two bugs:
1. `Skeleton.write()` returns another skeleton than the one that is
passed. Especially when working with subskels, this is an unwanted
behavior.
2. Due the previous full skeleton return, the `postSaveHandlers` of
`RelationalBones` are ALL executed, also when the bone is not part of
the skeleton. This yields in a query that is fired every time on every
bone, which slows down the entire process of the write() with nonsense
queries.