fix: `NumericBone` ignores precision on read & write (#1425)
This bug came in with the vi-admin 4.8.x which enforces client-side
validation, also of the precision.

In case there of this bone definition
```python
    timedrift = NumericBone(
        descr="Zeitkorrektur",
        defaultValue=0,
        indexed=False,
        precision=1
    )
```

And a computed value that was -4.03169999999, the client rejected to
submit the mask with

![image](https://github.com/user-attachments/assets/379e4bb1-2fba-4d1e-abfb-98503c087122)

The problem here is, that the client enforces everything right, but the
server loads the value -4.03169999999 which is not valid by its own
bonedefinition, so -4.0 would be the correct value.

This pull request fixes NumericBone to store and load only values with
the correct precision set.