fix: don't compute a bone in a not existing relation
If you have a compute bone in a skel, thats referenced by a
RelationalBone (e.g. OrderSkel -> CartNodeSkel in viur-shop) and the
referenced skel is deleted (e.g. CartNodeSkel) this skleton cannot
longer be `read` here:
```
cloned_skel = skeletonByKind(skel.kindName)()
cloned_skel.read(skel["key"])
```

But here's not handling if the `read` _fails_ and returns `None`,
therefore the skel has only the default values and the computing methods
fails if it expeect for example a not-`None` `key` in the skel.

Since I detected this on rebuild search index, this breaked the entire
task and even the default error logging of the skel failed, since it
coulnd't be unserialized.