fix: don't compute a bone in a not existing relation (#1443)
If you have a compute bone in a skel that's referenced by a
`RelationalBone` (e.g. `OrderSkel` --> `CartNodeSkel` in viur-shop), and
this referenced skel is deleted (e.g. `CartNodeSkel`), this skeleton
cannot longer be `read()` here:

```py
cloned_skel = skeletonByKind(skel.kindName)()
cloned_skel.read(skel["key"])
```

But here's not handling if the `read()` _fails_ and returns `None`,
therefore the skel still has only the default values and the computing
methods fails if it expect for example a not-`None` `key` in the skel.

Since I detected this on rebuild search index, this broke the entire
task and even the default error logging of the skel failed, because it
couldn't be unserialized.