fix: `processRemovedRelations` failes if the source entity has other relations set (#1468)
If a relation with the key `key1` has been deleted. In the skeleton
might be other `RelationalBone`s which have a `key2`. These obviously do
not match. However, this then leads to the `else` being reached and you
get the following error:

> In <viur.datastore.Key source_kind/key123, parent=None>, no handling
for relational_value={'dest': <SkeletonInstance of RefSkelForfile with
{'key': <viur.datastore.Key file/key2, parent=None>,...}>