feat: `ImageBone` (and some little splitting of skeleton.py) (#1454)
Adds the `ImageBone`, which is a pre-configured `FileBone` with the
ability to provide `alt`-texts for the image in different languages.
This is part of an initiative for accessibility-features in ViUR.

To achieve this, the module `skeleton.py` was required to be split up
into several modules for better integration, and to avoid circular
imports.

---------

Co-authored-by: agudermann <47318461+ArneGudermann@users.noreply.github.com>