fix: `User.onLogin()` shall be called earlier
This change would allow the User.onLogin()-hook to update things in the provided UserSkel and take this (probably updated) UserSkel into the session.
Otherwise, it might happen that the UserSkel is internally updated, but the old UserSkel remains in the session.

Here's an actual example for such a hook from a customer project (without code that does the `current.session.get()["user"]`-re-assignment and `current.user.set()`-call, which are necessary without this change):

```py
    def onLogin(self, skel):
        # Synchronize user with external system
        if not skel["manualoverride"] and skel["debitor"]:
            ok, _ = conf.main_app.thirdparty.sync_user_skel(skel)
            if ok:
                skel.write(update_relations=False)

                # Is user still active, or meanwhile flagged locked?
                if not self.is_active(skel):
                    session.killSessionByUser(skel["key"])
                    raise errors.Redirect("s/user_locked")

        # Super-Login
        super().onLogin(skel)
```