fix: `FileBone` with defined defaults
This PR fixes the case that happens when initializing a FileBone after system initialization (e.g. in an ad-hoc action skel).
In that case, the "dlkey" didn't become part of the _refSkelCache because this class was generated in RelationalBone.__init__() without the later added "dlkey".
"dlkey" is mandatory to this bone.